implement connection to mongodb

+++ resources/product.go
@@ -4,6 +4,6 @@ package resources
 type Product struct {
 	Id string `json:"Id"`
 	Name string `json:"Name"`
-	QuantityInStock int16 `json:"QuantityInStock"`
-	Price int16 `json:"Price"`
+	QuantityInStock int `json:"QuantityInStock"`
+	Price int `json:"Price"`
 }
